Move all the components for the chat page inside the chat directory to make it easier to navigate. - Linked issue: #100
