+++ soen341-frontend/app/chat/components/message-input.tsx
