+++ soen341-frontend/app/chat/components/message-item.tsx
