+++ soen341-frontend/app/chat/components/sidebar.tsx
