+++ soen341-frontend/app/chat/page.tsx
@@ -1,4 +1,4 @@
-import { Messaging } from "@/Components/messaging";
+import { Messaging } from "@/app/chat/components/messaging";
 
 export default function Chat() {
   return <Messaging />;
